backend/requirements.txt
flask>=2.0
flask-cors>=3.0
tensorflow>=2.10.0
pillow
numpy
scikit-learn
python-dateutil